Completely rewrote the config system.
Configuration data now resides in a Namespace.

A large number of attributes on TutorialApp which merely cached config
data have been removed.

The config variable on TutorialApp has been renamed cfg, to avoid clashes
with the tk attr .config (notwithstanding that TutorialApp is not a tk
subclass).

General refactoring and prettiness.