Added TutorialNodeVisitor.involved_identifiers method
It's often necessary to check if a student has used a particular
identifier in a particular context.

Depending on the type of node, this can be quite an involved process.
For example, a student might write extremely convoluted if statements,
where we expect only a simple one.

TutorialNodeVisitor.involved_identifiers will recursively determine all
identifiers which have been used in a statement.

So, for example, if the student writes this statement:

    if x > 3 and (y or z < 4):
        pass

then TutorialNodeVisitor.involved_identifiers in the context of the if
statement (ie visit_If) will return:

    ['x', 'y', 'z']

even though the tree itself is extremely complex.

The order of identifiers is undefined, but will broadly be depth-first.

This will be useful in implementing #15