Replaced run_student_code with run_in_student_context
StudentTestCase.run_in_student_context will take a function object and
execute it in the context of the student code.

This is intended to be used for obtaining test results.  The output of the
function will be returned.

Functions can contain arbitrary code, eg:

    def test_things(self):
        def _get_results(self):
	    first_result = dec2base(10, 10)
	    second_result = dec2base(9, 2)

	    return first_result, second_result

	result = self.run_in_student_context(_get_results)

Because functions are run in the context of the student code, they may
reference any object which is (or should be) defined in that code.  The
flip side of that is that they *cannot* reference objects defined in the
local scope (IRRESPECTIVE OF THE USUAL CLOSURE RULES).

See #25 for more discussion on this.

This commit also updates recursive_digits2.tut to use the new approach,
but not any of the other tutorial problems.