Added note regarding removing student_function_name
This should be considered a TODO on #25

Because we no longer extract the student function, there's actually no
need to know about its name.  It can therefore be removed from tester.py
(and I will do so soon).

However, keeping the variable in the config file may still be useful, as
it enables the useful anlysis result of 'you haven't actually defined that
function, silly'.

Unsure whether that's worth the added screwing around, though.