Completely redesigned the recursive visits system
Key points:
  * it is no longer necessary to use @visit_recursively
  * analysis.py must now define ._analyse(), not .analyse(text)

This change was made necessary by the addition of functionality in
TutorialNodeVisitor (the superclass).  For full details see the
descriptive comment in TutorialNodeVisitor.generic_visit, but the short
version is that recursing from the parent class caused issues with nodes
being visited more than once, or (under an alternative strategy) not at
all.