Partially refactored code/fontChooser.py
This was only a partial refactor (change of name, and move to gui module).

The file would defintely benefit from a full refactor, but it's quite a
large file and is fully self-contained.

I don't see the benefit of spending the time on cleaning it up at the
moment, especially considering the underlying functionality is very rarely
used by students.