TutorEditor now correctly uses Bindings from idlelib
Previously, bindings were defined in TutorBindings.py, and then used in
TutorEditor.py by calling initialise() and then accessing a global
variable, which was passed to fill_menus.

This isn't the way that idlelib itself handles things.  idlelib actually
reads off a Bindings class variable in the EditorWindow, and fills the
menus with that information in the constructor.  Because we were also
filling the menus in our constructor, we were getting duplicate menus.
That is no longer an issue.

However, we do have another issue: the existing code to special-case OS
X depended on idlelib's runtime detection of the OS.  Unfortunately,
because tk has not initialised at the time the TutorBindings module is
executed, we don't have access to that information.

There are a number of hacky ways around this.  For example, we could
replace Bindings with a class that dynamically builds the bindings using
__getattribute__ (which would be called at runtime, and would thus have
access to idlelib.macosxSupport).  That's probably the best bet.

For now, the special OS X stuff is disabled.  I'll fix it up once other
things are working.