Implemented automatic sync.
Closes #10.

Remaining work on things related to this: we had a discussion much
earlier about the possiblity of putting MPT on the EAIT image.  This
would save students from the abject horror of having to learn how to
install things.

The problem with that is that students might be tempted to put the
answers folder in a publicly-accessible location (which includes the
windows documents folder on an EAIT workstation, for some reason).

There's (at least) four ways we could deal with this:
  * we could introduce a heuristic which rejects any tutorial package
    which the system thinks is on an EAIT system and which appears to
    place the answers folder in a publicly-accessible location
      * problem: this is messy
      * problem: the heuristic could get it wrong
  * alternatively, we could use that same heuristic, but instead of
    attempting to validate the configuration settings, we could instead
    just delete every local answer after synchronisation on an EAIT
    workstation
      * this is much, much cleaner than the above
      * problem: the heuristic could still get it wrong
  * we could just delete every answer file every time the program quits
      * problem: if synchronisation fails, we either have to keep those
        nasty answers around, or delete student work
  * we could never store any data locally at all - changing problems
    could automatically upload the latest answer for the last problem,
    and download the latest answer for the one that's being switched to
      * this requires an always-on internet connection