Answer path components are now sanitised.
This introduces a dependency on werkzeug on the server side.
Only the util function secure_filename is used, however, so it would
theoretically be possible to extract that function if necessary (subject
to licensing issues).
Because this is a server dependency, not a client one, it's not a huge
problem; we can happily install the module on the server.

Closes #42.
However, #42 will need to be revisted when the submission code is updated
(which is necessary due to the lack of encryption).