Added web implementation of submission
This uses sha512 hashes to identify tutorials.

Note that this does pose a potential issue: if we update a tutorial
package, and push that update out to users, then the hash will change.

TODO: Write a script that can update existing student submissions to match
the new hash, OR one that will collate answers using any past hash of the
tutorial contents.