Added TODO note in create_tutorial.py about tutorial_hashes
This commit is really just to give me a chance to include the following
discussion, which I foolishly forgot to add to the previous commit before
pushing.

So why use sha512 hashes instead of UUIDs?
Well, we need to verify that the tutorial atually hasn't been changed.
Obviously we can't do this in a tamper-resistant way, given that MPT is
open source, but we have to try.

Because of this need for verification, we will be sending a hash to the
server in any case.  The server needs to be able to map that hash to a
tutorial.  Now, we could also send a UUID, and use that to look up the
valid hash.  But I don't see any advantage to that over just sending the
hash, and using that to map directly to the tutorial.

Of course, we do have that issue with changed tutorial packages, which
will change the hash.  But that can be resolved server-side, without the
client ever needing master IDs for each tutorial.