Re-added key bindings (with a caveat).
Caveat: bindings do not appear to work until a menu is opened and closed
(even if nothing is clicked).  My guess is that, because of the login
popup, the main window (and thus Tk) doesn't have focus until then.

Also, TutorEditor needs to reset the bindings that it's interested in.
An attempt to use Tk.bind_all to grab them app-wide, then reject the
unnecessary ones in EditorWindow (which would have been much neater from a
design perspective) interfered with something.
In short, the <Key> event on EditorWindow.text was only grabbing every
second key or so, and so couldn't reliably break the event chain.
God knows why.  Stupid Tk.