Removed _current_class
This was causing confusion, due to the way that .peek works.

The _current_class would only be something other than None if at *class
scope*; if inside a method, it would return None.

I've removed this for now, rather than change it to return the class, as I
think it's better for it to be consistent with _current_function (which
will give a fully qualified name).