Fixed bug when execing support.py
This was another incarnation of the globals/locals issue (see comment in
exec_module for a full description).

Essentially, we need to only pass a single dict to exec wherever
possible.  Doing otherwise invites python to make incorrect assumptions
about how module-level references are used.

As a result, the testing framework now only uses a single dict (for both
locals and globals) from the very beginning.