Added requirement for command to gui tutes (oops)
Note that the print statement currently goes to console, not (as might
be expected) to the textbox.

This is because stdout is only redirected during testing.

It might be necessary to redirect stdout more generally, but I'm a bit
uncomfortable with that, as I still want error messages to go to
console, and it's a bit inconsistent to redirect one and not the other.