Messagebox workaround for OS X.  Closes #49
The solution is to intercept all messagebox calls and call both update
and update_idletasks on the rook window after each messagebox has
appeared.

We get a reference to the root window in a really hacky way: popup a
Toplevel, read off its master, and then destroy it immediately.

Oh well, it works :)