Fixes infinite recursion in input (closes #68)
Infinite recursion was occurring because the module builtins (from
'import builtins') is actually the same as gbls['__builtins__']['input']
in most cases; because of this, recursing on builtins.input() in
glbs['__builtins__]['input'] is thus an infinite loop.

Also cleaned up context manager to correctly preference replacing
gbls['input'] to gbls['__builtins__']['input'].  This is necessary
because it is the former which will be called by test code.