Hotfix for #69 (default font selection)
Font now defaults to tkinter default, which is readable on all
operating systems.  (Tk doesn't seem to like retina displays, so it's
arguably a bit small on OS X still.)

Not closing yet, as we need to come up with italicised/bold etc
versions, as well as use the default fixed-width font.  That's not
working at the moment because the TutorialFrame calls get_configs before
mainloop is entered, meaning that the font info is unavailable.
(Tkinter really does love using Tk() as a singleton for *everything*)

TODO:
* change get_configs to use tkinter.font
* handle special fonts (eg TkDefaultFont, TkFixedFont), noting that
  tkinter.font.nametofont returns a singleton which can't be mutated
  without breaking the rest of the app