Added color-coding to menus
Only works on child menus atm.
Not tested on Windows or Linux (lack of computers due to moving).

See #85

Doesn't close the issue yet; top-level highlighting is a bit more
fiddly (leads to ugly code), and I'm considering going without, but I'd
like to see the results on Windows first.