Removed font change feature.
This closes #48.

I don't believe that the feature is useful enough to justify inclusion,
especially as it would require a lot of introspection to work correctly
(eg, working out which fonts are fixed or not).

All UI widgets which should use a fixed-width font will now use the
appropriate one for the OS in question.