Now only updates to more recent tutorials
Previously, the server version would always be preferred.
This was causing problems in development, as the server's version of the
tutorials would be synchronised over any local changes.

As a consequence, it's no longer possible to revert to an older version
on the server and push that to students without changing the timestamp.