Fixed missing popup on login failure.
As a side-effect, WebAPI is no longer aware of whether the login was a
failure or if the user cancelled.

At the moment, that's ok (even though it goes against what we said for
issue #63), because a bunch of UI code is tied up in SessionManager
anyway (and so the user will still see a popup).

It would be preferable to move all of that code to TutorialApp (via
WebAPI), but that's a decent undertaking so I'm going to put it off for
now.