Better reporting of allow_late flag; ability to unset late flag.
Previously, parse_submission_log would not give information about late flags for unsubmitted problems.
Added logging of the username/timestamp of who performs admin actions on a user's account in support.py.
Unsetting late flag adds a new entry to the log (this is just because adding a line to a file is easier than removing one - this functionality wouldn't be necessary if the server storage is moved to a database)