Interpreter now works on windows
Closes #120
The root cause of this issue was twofold.
First, windows doesn't include python in $PATH by default.
Second, there's no simlink to idle in /Scripts on windows.

As a result, on win32 MPT will now exec the idle script in the python
interpreter.  The locaion of the script is determined relative to the
interpreter which is running MPT.