Now resizes properly on small screens.
MPT no longer tracks explicit heights.  Default heights for the
testing/analysis outputs are fixed at 5.

The TutorialFrame (which renders the HTML description) now has a nominal
height of 0, meaning that it will be shrunk first by the geometry
manager when the window is resized.

The TutorialApp will now request a fixed geometry, but never more than
the screen size.  This should mean that it is created at the correct
size on small monitors (or, at least, close enough).

Closes #109.