Installer should now rm itself upon bootstrapping
This should make it less likely that students keep running the installer
over and over again to launch MPT.

I don't have try/except on OSError guarding the remove call.  In theory,
this should be ok; it's definitely not an error to remove a running
script on Unix, but it may be on Windows.  No problems were encountered
with this behaviour on Windows 7, at least in a simple test script.
This should be verified in the lab environment.