Added batch file to run in Anaconda's python exe
This is a stopgap measure to enable students on the lab machines to run
MyPyTutor without having to open a command prompt.

Instead, all they have to do is select 'Open With', and navigate to the
batch file.  It will then pass on all arguments to the Python script.
(It's not possible for students to directly choose to open with the
actual Python executable, as Windows refuses to have two entries in that
context menu with the same filename.  Seriously.  Such crap.)

I probably should say that it will pass on all arguments *in theory*.
It's entirely possible that Windows will apply its usual crap handling
of command line arguments (in particular, an odd approach to quoting)
and that this may cause certain things to break.  You have been warned.