Fixed incorrect detection of hashability
This bug occurred when a type as a whole was hashable, but certain
instances of that type could not be.  By far the most common example of
that is a tuple, eg (1, [2, 3, 4]) is not hashable, but (1, 2) is.

We now simple try/except on hash to determine if a type is hashable.

An alternative approach to determining hashability, by explicitly
recursing on iterable objects, proved to be a very bad idea in practice,
as infinite loops could result if an object was self-referential.

Closes #151