Replaces input during compilation
This fixes timeouts caused by calls to input() at global scope

For some reason, the Alarm class (which uses thread._interrupt_main) was
not getting an opportunity to catch these.
My guess would be that input() on sys.stdin operates as blocking IO at
the kernel level and so cannot be interrupted by a signal.

Closes #155